+++ ema-examples/src/Ema/Example/Ex06_Markdown.hs
@@ -19,7 +19,7 @@ import Ema.CLI qualified as CLI
 import Ema.Route.Generic.TH
 import Ema.Route.Lib.Extra.PandocRoute qualified as Pandoc
 import Ema.Server (EmaServerOptions (..))
-import Ema.Server.WebSocket (EmaWsHandler (..), wsClientJS)
+import Ema.Server.WebSocket.Options (EmaWsHandler (..), wsClientJS)
 import Network.WebSockets qualified as WS
 import Optics.Core ((%))
 import System.Directory (makeAbsolute)