+++ ema/src/Ema/Server/WebSocket.hs
@@ -1,10 +1,6 @@
-{-# LANGUAGE QuasiQuotes #-}
-{-# LANGUAGE RecordWildCards #-}
-
 module Ema.Server.WebSocket where
 
 import Control.Monad.Logger
-import Data.Default (Default (def))
 import Data.LVar (LVar)
 import Data.LVar qualified as LVar
 import Ema.Asset (
@@ -16,46 +12,14 @@ import Ema.Route.Prism (
   fromPrism_,
  )
 import Ema.Server.Common
+import Ema.Server.WebSocket.Options (EmaWsHandler (..))
 import Ema.Site (EmaStaticSite)
-import NeatInterpolation (text)
 import Network.WebSockets (ConnectionException)
 import Network.WebSockets qualified as WS
 import Optics.Core (review)
 import Text.Printf (printf)
 import UnliftIO.Async (race)
 import UnliftIO.Exception (try)
-
-{- | A handler takes a websocket connection and the current model and then watches
-   for websocket messages. It must return a new route to watch (after that, the
-   returned route's HTML will be sent back to the client).
-
-  Note that this is usually a long-running thread that waits for the client's
-  messages. But you can also use it to implement custom server actions, by handling
-  the incoming websocket messages or other IO events in any way you like.
-
-  Also note that whenever the model is updated, the handler action will be
-  stopped and then restarted with the new model as argument.
--}
-newtype EmaWsHandler r = EmaWsHandler
-  { unEmaWsHandler :: WS.Connection -> RouteModel r -> LoggingT IO Text
-  }
-
-instance Default (EmaWsHandler r) where
-  def = EmaWsHandler $ \conn _model -> do
-    msg :: Text <- liftIO $ WS.receiveData conn
-    log LevelDebug $ "<~~ " <> show msg
-    pure msg
-    where
-      log lvl (t :: Text) = logWithoutLoc "ema.ws" lvl t
-
-data EmaServerOptions r = EmaServerOptions
-  { emaServerShim :: LByteString
-  , emaServerWsHandler :: EmaWsHandler r
-  }
-
-instance Default (EmaServerOptions r) where
-  def =
-    EmaServerOptions wsClientJS def
 
 wsApp ::
   forall r.
@@ -120,17 +84,3 @@ wsApp logger model emaWsHandler pendingConn = do
           _ ->
             log LevelError $ "Websocket error: " <> show connExc
         LVar.removeListener model subId
-
--- Browser-side JavaScript code for interacting with the Haskell server
-wsClientJS :: LByteString
-wsClientJS =
-  encodeUtf8
-    [text|
-        <script type="module" src="https://cdn.jsdelivr.net/npm/morphdom@2.6.1/dist/morphdom-umd.min.js"></script>
-
-        <script type="module">
-        ${wsClientJSShim}
-        
-        window.onpageshow = function () { init(false) };
-        </script>
-    |]