+++ ema/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Relax `base` constraint forever
 - Live server shim/websocket customization ([\#152](https://github.com/srid/ema/pull/152)) @lucasvreis
 - `Ema.CLI`: The `Action` type is no longer a GADT.
+- `Ema.Server`: This module has be split into several smaller modules
 
 ## 0.10.2.0 (2023-08-09)